requirements.txt
streamlit>=1.30.0
tensorflow>=2.12.0
numpy>=1.23.0
pandas>=1.5.0
matplotlib>=3.6.0
scikit-learn>=1.2.0
fpdf2>=2.7.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
